+++ firmware/Azhoo/IncrementalEncoderUno.cpp
@@ -1,8 +1,8 @@
 
 #include "IncrementalEncoderUno.h"
         
-void IncrementalEncoderUno::init(int32_t pulses_per_revolution,
-                                 int32_t increment, 
+void IncrementalEncoderUno::init(uint16_t pulses_per_revolution,
+                                 int8_t increment, 
                                  uint8_t pin_a,
                                  uint8_t pin_b,
                                  EncoderInterruptHandler callback)
